Use config macro for Buffer Allocation Method
