+++ source/include/FreeRTOS_IP_Private.h
@@ -337,10 +337,6 @@ typedef union xUDPPacketHeader
 } UDPPacketHeader_t;
 extern UDPPacketHeader_t xDefaultPartUDPPacketHeader;
 
-
-/* True when BufferAllocation_1.c was included, false for BufferAllocation_2.c */
-extern const BaseType_t xBufferAllocFixedSize;
-
 /* As FreeRTOS_Routing is included later, use forward declarations
  * of the two structs. */
 struct xNetworkEndPoint;
