+++ test/unit-test/FreeRTOS_DNS_Networking/FreeRTOS_UDP_IP_stubs.c
@@ -41,8 +41,6 @@
 #include "FreeRTOS_IP_Private.h"
 
 
-const BaseType_t xBufferAllocFixedSize = pdTRUE;
-
 void vPortEnterCritical( void )
 {
 }
