+++ test/unit-test/FreeRTOS_DNS_Parser/FreeRTOS_DNS_Parser_utest.c
@@ -78,8 +78,6 @@ static void dns_callback( const char * pcName,
 
 /* ===========================  EXTERN VARIABLES  =========================== */
 
-extern BaseType_t xBufferAllocFixedSize;
-
 extern pucAddrBuffer[ 2 ];
 extern pucSockAddrBuffer[ 1 ];
 
@@ -90,7 +88,6 @@ extern pucSockAddrBuffer[ 1 ];
  */
 void setUp( void )
 {
-    xBufferAllocFixedSize = pdFALSE;
     xDefaultPartUDPPacketHeader.ucBytes[ 0 ] = 1;
     xDefaultPartUDPPacketHeader.ucBytes[ 1 ] = 2;
     xDefaultPartUDPPacketHeader.ucBytes[ 2 ] = 3;
@@ -995,7 +992,6 @@ void test_DNS_TreatNBNS_Fail_BufferAllocation1( void )
     pxNetworkBuffer.xDataLength = 3000;
 
     hook_return = pdTRUE;
-    xBufferAllocFixedSize = pdTRUE;
     pxUDPPayloadBuffer_to_NetworkBuffer_ExpectAnyArgsAndReturn( &pxNetworkBuffer );
     uxIPHeaderSizePacket_IgnoreAndReturn( ipSIZE_OF_IPv4_HEADER );
     usChar2u16_ExpectAnyArgsAndReturn( dnsNBNS_FLAGS_OPCODE_QUERY ); /* usFlags */
@@ -1025,7 +1021,6 @@ void test_DNS_TreatNBNS_success_BufferAllocation1( void )
     pxNetworkBuffer.xDataLength = 300;
 
     hook_return = pdTRUE;
-    xBufferAllocFixedSize = pdTRUE;
     pxUDPPayloadBuffer_to_NetworkBuffer_ExpectAnyArgsAndReturn( &pxNetworkBuffer );
     uxIPHeaderSizePacket_IgnoreAndReturn( ipSIZE_OF_IPv4_HEADER );
     usChar2u16_ExpectAnyArgsAndReturn( dnsNBNS_FLAGS_OPCODE_QUERY ); /* usFlags */
@@ -1111,7 +1106,6 @@ void test_DNS_TreatNBNS_success_nbns_non_fixed_size_buffer2( void )
     size_t uxBufferLength = 300;
     uint32_t ulIPAddress;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t pxNetworkBuffer = { 0 };
     struct xNetworkEndPoint xEndPoint = { 0 };
     pxNetworkBuffer.pxEndPoint = &xEndPoint;
@@ -1508,7 +1502,6 @@ void test_DNS_ParseDNSReply_answer_lmmnr_reply_xBufferAllocFixedsize( void )
     struct freertos_addrinfo * pxAddressInfo;
     uint16_t usPort;
 
-    xBufferAllocFixedSize = pdTRUE;
     uint8_t * nullAddress = NULL;
     NetworkEndPoint_t xEndPoint = { 0 };
 
@@ -2562,7 +2555,6 @@ void test_DNS_ParseDNSReply_answer_lmmnr_reply_valid_new_netbuffer3( void )
     usChar2u16_ExpectAnyArgsAndReturn( dnsTYPE_AAAA_HOST ); /* usType */
     usChar2u16_ExpectAnyArgsAndReturn( dnsCLASS_IN );       /* usClass */
     hook_return = pdTRUE;
-    xBufferAllocFixedSize = pdTRUE;
     pxUDPPayloadBuffer_to_NetworkBuffer_ExpectAnyArgsAndReturn( &pxNetworkBuffer );
     FreeRTOS_FindEndPointOnNetMask_ExpectAndReturn( xIPPacket->xIPHeader.ulSourceIPAddress, 6, &xEndPoint );
     usGenerateChecksum_ExpectAnyArgsAndReturn( 555 );
