+++ test/unit-test/FreeRTOS_TCP_Transmission/FreeRTOS_TCP_Transmission_utest.c
@@ -352,7 +352,6 @@ void test_prvTCPSendRepeated_Zero_To_Send( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = ( StreamBuffer_t * ) 0x12345678;
     pxSocket->u.xTCP.eTCPState = eCONNECT_SYN;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1202,8 +1201,6 @@ void test_prvTCPBufferResize_Fixed_Size_With_Buffer( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdTRUE;
-
     uxIPHeaderSizeSocket_ExpectAnyArgsAndReturn( ipSIZE_OF_IPv4_HEADER );
 
     pReturn = prvTCPBufferResize( pxSocket, pxNetworkBuffer, 500, 0 );
@@ -1219,8 +1216,6 @@ void test_prvTCPBufferResize_Fixed_Size_Without_Buffer( void )
     pxSocket = &xSocket;
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
-
-    xBufferAllocFixedSize = pdTRUE;
 
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
@@ -1240,7 +1235,6 @@ void test_prvTCPBufferResize_Without_Buffer( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
 
@@ -1261,7 +1255,6 @@ void test_prvTCPBufferResize_Without_Buffer_Null_New_Buffer( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
 
@@ -1282,7 +1275,6 @@ void test_prvTCPBufferResize_With_Buffer_LT_Needed_GT_Last_Packet( void )
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
     pxNetworkBuffer->xDataLength = 500;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
 
@@ -1305,7 +1297,6 @@ void test_prvTCPBufferResize_With_Buffer_LT_Needed_LT_Last_Packet( void )
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
     pxNetworkBuffer->xDataLength = 10;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
 
@@ -1327,7 +1318,6 @@ void test_prvTCPBufferResize_With_Buffer_GT_Needed( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ucEthernetBuffer;
 
@@ -1392,7 +1382,6 @@ void test_prvTCPPrepareSend_State_Syn_Zero_Data( void )
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = ( StreamBuffer_t * ) 0x12345678;
     pxSocket->u.xTCP.eTCPState = eCONNECT_SYN;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1417,7 +1406,6 @@ void test_prvTCPPrepareSend_State_Syn_Zero_Data_Win_Change( void )
 
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = ( StreamBuffer_t * ) 0x12345678;
     pxSocket->u.xTCP.eTCPState = eCONNECT_SYN;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1441,7 +1429,6 @@ void test_prvTCPPrepareSend_State_Syn_Zero_Data_Keep_Alive( void )
 
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = ( StreamBuffer_t * ) 0x12345678;
     pxSocket->u.xTCP.eTCPState = eCONNECT_SYN;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1466,7 +1453,6 @@ void test_prvTCPPrepareSend_State_Established_Zero_Data_KLCount1_Age_GT_Max( voi
 
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = NULL;
     pxSocket->u.xTCP.eTCPState = eESTABLISHED;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1494,7 +1480,6 @@ void test_prvTCPPrepareSend_State_Established_Null_Buffer_Zero_Data_KLCount0_Age
     pxSocket = &xSocket;
     pxNetworkBuffer = NULL;
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = NULL;
     pxSocket->u.xTCP.eTCPState = eESTABLISHED;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1521,7 +1506,6 @@ void test_prvTCPPrepareSend_State_Established_Null_Buffer_Zero_Data_KLCount1_Age
     pxSocket = &xSocket;
     pxNetworkBuffer = NULL;
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = NULL;
     pxSocket->u.xTCP.eTCPState = eESTABLISHED;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1552,7 +1536,6 @@ void test_prvTCPPrepareSend_State_Established_Zero_Data_KLCount1_Age_GT_Max_Win_
 
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = NULL;
     pxSocket->u.xTCP.eTCPState = eESTABLISHED;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1585,7 +1568,6 @@ void test_prvTCPPrepareSend_State_Established_Non_Zero_Data_No_Buffer( void )
 
     StreamBuffer_t StreamBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     pxSocket->u.xTCP.txStream = ( StreamBuffer_t * ) &StreamBuffer;
     pxSocket->u.xTCP.eTCPState = eESTABLISHED;
     pxSocket->u.xTCP.usMSS = 1000;
@@ -1612,7 +1594,6 @@ void test_prvTCPPrepareSend_State_Established_Non_Zero_Data_MSS_0_KLCount4( void
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ReturnEthernetBuffer;
@@ -1645,7 +1626,6 @@ void test_prvTCPPrepareSend_State_Established_Non_Zero_Data_Not_Close_Not_ShutDo
     pxNetworkBuffer = &xNetworkBuffer;
     pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
 
-    xBufferAllocFixedSize = pdFALSE;
     uint8_t ReturnEthernetBuffer[ ipconfigNETWORK_MTU ];
     NetworkBufferDescriptor_t NewNetworkBuffer;
     NewNetworkBuffer.pucEthernetBuffer = ReturnEthernetBuffer;
